Metric with the peak resident memory usage of a process for each query.
For that, we use VmHWM (Peak resident set size ("high water mark"))
from /proc/PID/status, and reset it between queries with echo 5 >
/proc/PID/clear_refs.
It's a bit tricky to do that when the process is running inside a
docker container, as then /proc/PID/clear_refs is only writeable by
root. We therefore need to write to clear_refs inside the docker
container, which requires finding the PID of the process inside the
docker container PID namespace.

Also, this metric should be taken with a grain of salt: it might not
always be what you expect as allocator typically don't release
previously allocated memory to the OS as soon as the memory is no
longer needed.